feat: update delete task button to show confirmation modal
This commit updates the delete task button to show the
confirmation modal when clicked. The modal will ask the
user to confirm that they want to delete the task. If the
user confirms, the task will be deleted. If the user cancels,
the modal will close and the task will not be deleted.